gvisor: ignore OOM panics
OSS-Fuzz is fuzzing deserialization code, where objects are prepended
with headers that includes the object lengths. When one of those lengths
is too large, malloc fails and causes a panic. This is what we want to
happen.

This change works around this by having the test launch a subprocess.
The subprocess runs the actual fuzzing test, and the parent passes along
the subprocess return value unless it sees an OOM.

Fixes #7826.